feat: Rest of Chapter 20 Java Interoperability
JvmAnnotations to magically make the wonders of Kotlin available to Java
code. Cool. But easier to just write in one or the other. Should be easy
enough to add a little Kotlin to Java, or a little Java to Kotlin, but
between there - half and half, for example - could be a real pain...